ongoing/requirements.txt
adafruit-ampy==1.0.7
appdirs==1.4.3
apt-clone==0.2.1
apturl==0.5.2
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.6.6
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cs50==5.0.3
cupshelpers==1.0
DateTime==4.3
defer==1.0.6
distlib==0.3.0
docker==4.2.0
docker-pycreds==0.4.0
dominate==2.4.0
ecdsa==0.15
esptool==2.8
filelock==3.0.12
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Moment==0.9.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gigantum==0.19
httplib2==0.17.0
idna==2.9
importlib-metadata==1.5.0
importlib-resources==1.2.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.3.1
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
nemo-emblems==4.4.0
netifaces==0.10.9
onboard==1.4.1
packaging==20.3
PAM==0.4.2
pexpect==4.8.0
Pillow==7.0.0
pluggy==0.13.1
protobuf==3.11.3
psutil==5.4.2
ptyprocess==0.6.0
py==1.8.1
pyaes==1.6.1
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pygobject==3.26.1
PyICU==1.9.8
pyinotify==0.9.6
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyparted==3.11.1
PyQt5==5.14.1
PyQt5-sip==12.7.1
pyRFC3339==1.1
pyserial==3.4
pytest==5.3.5
python-apt==1.6.5+ubuntu0.2
python-debian==0.1.36
python-dotenv==0.12.0
python-xapp==1.8.1
python-xlib==0.26
pytz==2019.3
pyxdg==0.26
PyYAML==5.3
reportlab==3.5.34
requests==2.23.0
requests-unixsocket==0.2.0
safeeyes==2.0.9
sessioninstaller==0.0.0
setproctitle==1.1.10
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.13
sqlparse==0.3.1
system-service==0.3
systemd-python==234
termcolor==1.1.0
tinycss==0.4
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.8
vboxapi==1.0
virtualenv==20.0.8
visitor==0.1.3
wcwidth==0.1.8
websocket-client==0.57.0
Werkzeug==1.0.0
wrapt==1.12.0
xkit==0.0.0
youtube-dl==2020.3.6
zipp==3.1.0
zope.interface==4.7.1
